requirements.txt
python-telegram-bot==21.9
newspaper3k==0.2.8
trafilatura==1.8.0
fastapi==0.115.0
uvicorn==0.32.0
httpx>=0.27.0
python-dotenv==1.0.0
